Changed stamina recovery method

+++ src/engine/DrawManager.java
@@ -368,9 +368,9 @@ public final class DrawManager {
 			drawEntity(dummyShip, 40 + 35 * i, 10);
 	}
 
-	public void drawLivesbar(final Screen screen, final int lives) {
+	public void drawLivesbar(final Screen screen, final double lives) {
 		// Calculate the fill ratio based on the number of lives (assuming a maximum of 3 lives).
-		double fillRatio = (double) lives / 3.0;
+		double fillRatio = lives / 3.0;
 
 		// Determine the width of the filled portion of the rectangle.
 		int filledWidth = (int) (120 * fillRatio);
@@ -741,7 +741,7 @@ public final class DrawManager {
 	 *                       If the score is a new high score.
 	 */
 	public void drawResults(final Screen screen, final int score,
-			final int livesRemaining, final int shipsDestroyed, final int difficulty,
+			final double livesRemaining, final int shipsDestroyed, final int difficulty,
 			final float accuracy, final boolean isNewRecord) {
 		String scoreString = String.format("score %04d", score);
 		String difficultyString = "Difficulty ";
